remove readIntoBuffers from JSONHelper.cpp and move to .h file to fix the compiling issue.
